Navbar and footer are now in _app.js
Content rendered inside _app.js is persistent. That is, because the
navbar and the footer are always the same, we don't want to render it
from scratch every time we change to a new page. So I moved the
rendering of the navbar and the footer from layout.js to _app.js.